+++ src/transformers/models/sew/modeling_sew.py
@@ -249,7 +249,7 @@ def eager_attn_forward(
     if attention_mask is not None:
         attn_weights = attn_weights + attention_mask
 
-    attn_weights = nn.functional.softmax(attn_weights, dim=-1, dtype=torch.float32).to(query.dtype)
+    attn_weights = nn.functional.softmax(attn_weights, dim=-1)
 
     if layer_head_mask is not None:
         attn_weights = attn_weights * layer_head_mask.view(1, -1, 1, 1)
@@ -637,7 +637,6 @@ class SEWPreTrainedModel(PreTrainedModel):
         return attention_mask
 
 
-# Copied from transformers.models.wav2vec2.modeling_wav2vec2._compute_mask_indices
 def _compute_mask_indices(
     shape: Tuple[int, int],
     mask_prob: float,
